// src/auth/permission.enum.ts
export enum Permission {

  INSTITUTE_VIEW = 'INSTITUTE_VIEW',
  INSTITUTE_CREATE = 'INSTITUTE_CREATE',
  INSTITUTE_UPDATE = 'INSTITUTE_UPDATE',
  INSTITUTE_DELETE = 'INSTITUTE_DELETE',
  
  INVENTORY_VIEW = 'INVENTORY_VIEW',
  INVENTORY_CREATE = 'INVENTORY_CREATE',
  INVENTORY_UPDATE = 'INVENTORY_UPDATE',
  INVENTORY_DELETE = 'INVENTORY_DELETE',

  EQUIPMENT_VIEW = 'EQUIPMENT_VIEW',
  EQUIPMENT_CREATE = 'EQUIPMENT_CREATE',
  EQUIPMENT_UPDATE = 'EQUIPMENT_UPDATE',
  EQUIPMENT_ASSIGN = 'EQUIPMENT_ASSIGN',
  EQUIPMENT_DISPOSE = 'EQUIPMENT_DISPOSE',
  EQUIPMENT_DELETE = 'EQUIPMENT_DELETE',

  REPAIR_REQUEST_VIEW = 'REPAIR_REQUEST_VIEW',
  REPAIR_REQUEST_CREATE = 'REPAIR_REQUEST_CREATE',
  REPAIR_REQUEST_UPDATE = 'REPAIR_REQUEST_UPDATE',

  WORK_ORDER_VIEW = 'WORK_ORDER_VIEW',
  WORK_ORDER_ASSIGN = 'WORK_ORDER_ASSIGN',
  WORK_ORDER_COMPLETE = 'WORK_ORDER_COMPLETE',

  PROCUREMENT_VIEW = 'PROCUREMENT_VIEW',
  PROCUREMENT_CREATE = 'PROCUREMENT_CREATE',
  PROCUREMENT_APPROVE = 'PROCUREMENT_APPROVE',

  USER_VIEW = 'USER_VIEW',
  USER_CREATE = 'USER_CREATE',
  USER_UPDATE = 'USER_UPDATE',
  USER_DELETE = 'USER_DELETE',
  
  PERMISSION_VIEW = 'PERMISSION_VIEW',
  PERMISSION_CREATE = 'PERMISSION_CREATE',
  PERMISSION_TEMPORARY_CREATE = 'PERMISSION_TEMPORARY_CREATE',
  PERMISSION_REMOVE = 'PERMISSION_REMOVE',


  AUDIT_VIEW = 'AUDIT_VIEW'
}